Extract extract_progress to modules/logscan_progress_engine
Sprint 3, PR #8.  Pulls the 640-line extract_progress parser out
of LogscanAnalyzer into its own module.

Also extracts three small helpers that only extract_progress uses:
- _strip_divider_wrappers  -> strip_divider_wrappers(analyzer, msg)
- _extract_mapping_library -> extract_mapping_library(msg)
- _map_section_to_phase    -> map_section_to_phase(name)

Public API in modules/logscan_progress_engine.py:
- extract_progress(analyzer, content, ...)
    Long-running stateful parser that walks the log line-by-line
    tracking library statuses, phase timing, playlist runtime,
    and finished-run detection.  Returns a snapshot dict for
    Quickstart's live-progress dashboard.
- strip_divider_wrappers(analyzer, message)
- extract_mapping_library(message)
- map_section_to_phase(section_name)

The wrappers on LogscanAnalyzer preserve the existing method
surface (extract_progress, _strip_divider_wrappers,
_extract_mapping_library, _map_section_to_phase) so existing
tests and quickstart.py callers see zero API change.

This is a PURE MOVE.  No parser logic changed, no branches
reordered.  Every progress-payload field is byte-identical.

State the helper still reads from analyzer:
- global_divider  (via analyzer.set_global_divider /
                   analyzer.remove_repeated_dividers)

And methods still called through analyzer wrappers:
- extract_section_runtimes
- _normalize_library_name, _match_library_name
- _parse_hms_to_seconds

Impact:
- modules/logscan.py: 1317 -> 677 (-640 / -48.6%)
- modules/logscan_progress_engine.py: NEW 770 lines

Cumulative on logscan.py: 3291 -> 677 (-2614 total, -79.4%).
Just 77 lines above the 600-line target.

All 1284 tests pass.  Ruff + black clean.
